rename compiler to pkg
