+++ cmd/sleipnir/main.go
@@ -4,7 +4,7 @@ Copyright © 2024 NAME HERE <EMAIL ADDRESS>
 package main
 
 import (
-	"github.com/Ygg-Drasill/Sleipnir/compiler/cmd"
+	"github.com/Ygg-Drasill/Sleipnir/pkg/cmd"
 )
 
 func main() {
