--- compiler/errorHandler.go
@@ -1,1 +1,0 @@
-package main
