--- compiler/synthesis/optimizer.go
@@ -1,1 +1,0 @@
-package synthesis
