+++ pkg/analysis/lexer/keywords.go
