+++ pkg/analysis/lexer/token.go
