+++ pkg/analysis/utils/utils.go
