+++ pkg/synthesis/optimizer.go
@@ -1,0 +1,1 @@
+package synthesis
